moductypes: Add types
This is an experiment. I want the usual Python guarantee that
you can't just footgun yourself by passing the wrong thing through
a pointer argument, at least as much as is practical.

But a uctypes struct doesn't have a type of its own, like Point
or Rect; you can only check that the *size* matches some
underlying expected size.

This change allows a type like Rect to be defined in the core,
and of course if it's a distinct Python type it can be checked
in the core as well.

Restore the unix tests so we can test the added core functionality.

Signed-off-by: Jeff Epler <jepler@gmail.com>